+++ components/collector/src/source_collectors/azure_devops/base.py
@@ -51,7 +51,7 @@ class AzureDevopsJobs(SourceCollector):
         return URL(f"{await super()._api_url()}/_build")
 
     async def _parse_entities(self, responses: SourceResponses) -> Entities:
-        """Override to parse the jobs/pipelines."""
+        """Override to parse the jobs."""
         entities = Entities()
         for job in (await responses[0].json())["value"]:
             if not self._include_job(job):
@@ -88,3 +88,52 @@ class AzureDevopsJobs(SourceCollector):
     def __job_name(job: Job) -> str:
         """Return the job name."""
         return "/".join(job["path"].strip(r"\\").split(r"\\") + [job["name"]]).strip("/")
+
+
+class AzureDevopsPipelines(SourceCollector):
+    """Base class for pipeline collectors."""
+
+    async def _api_url(self, pipeline_id: int = None) -> URL:
+        """Extend to add the pipelines API path."""
+        pipeline_id_runs = f"/{pipeline_id}/runs" if pipeline_id else ""
+        # currently the pipelines api is not available in any version which is not a -preview version
+        return URL(f"{await super()._api_url()}/_apis/pipelines{pipeline_id_runs}?api-version=6.0-preview.1")
+
+    async def _active_pipeline_ids(self) -> list[int]:
+        """Find all active pipeline ids to traverse."""
+        api_pipelines_url = await self._api_url()
+        pipelines = (await (await super()._get_source_responses(api_pipelines_url))[0].json())["value"]
+        return [pipeline['id'] for pipeline in pipelines if 'id' in pipeline and self._include_pipeline(pipeline)]
+
+    async def _parse_entities(self, responses: SourceResponses) -> Entities:  # skipcq: PYL-W0613
+        """Override to parse the pipelines."""
+        entities = Entities()
+
+        for pipeline_id in await self._active_pipeline_ids():
+            api_pipelines_url = await self._api_url(pipeline_id)
+
+            for pipeline_run in (await (await super()._get_source_responses(api_pipelines_url))[0].json())["value"]:
+                if not self._include_pipeline_run(pipeline_run):
+                    continue
+
+                entities.append(
+                    Entity(
+                        key="-".join([str(pipeline_id), pipeline_run["name"]]),
+                        name=pipeline_run["name"],
+                        url=pipeline_run["_links"]["web"]["href"],
+                        build_date=str(parse(pipeline_run["finishedDate"]).date()),
+                        build_status=pipeline_run["state"]
+                    )
+                )
+        return entities
+
+    def _include_pipeline(self, job: Job) -> bool:
+        """Return whether this pipeline should be included."""
+        jobs_to_include = self._parameter("jobs_to_include")
+        if len(jobs_to_include) > 0 and not match_string_or_regular_expression(job["name"], jobs_to_include):
+            return False
+        return not match_string_or_regular_expression(job["name"], self._parameter("jobs_to_ignore"))
+
+    def _include_pipeline_run(self, job: Job) -> bool:  # skipcq: PYL-R0201
+        """Return whether this pipeline run should be included."""
+        return bool(job.get("finishedDate"))