+++ components/collector/src/source_collectors/jenkins/base.py
@@ -31,18 +31,18 @@ class JenkinsJobs(SourceCollector):
                     if self._build_datetime(job) > datetime.min
                     else "",
                 )
-                for job in self.__jobs((await responses[0].json())["jobs"])
+                for job in self._jobs((await responses[0].json())["jobs"])
             ]
         )
 
-    def __jobs(self, jobs: Jobs, parent_job_name: str = "") -> Iterator[Job]:
+    def _jobs(self, jobs: Jobs, parent_job_name: str = "") -> Iterator[Job]:
         """Recursively return the jobs and their child jobs that need to be counted for the metric."""
         for job in jobs:
             if parent_job_name:
                 job["name"] = f"{parent_job_name}/{job['name']}"
             if job.get("buildable") and self._include_job(job):
                 yield job
-            for child_job in self.__jobs(job.get("jobs", []), parent_job_name=job["name"]):
+            for child_job in self._jobs(job.get("jobs", []), parent_job_name=job["name"]):
                 yield child_job
 
     def _include_job(self, job: Job) -> bool:
@@ -66,5 +66,5 @@ class JenkinsJobs(SourceCollector):
         return "Not built"
 
     def _include_build(self, build) -> bool:  # pylint: disable=unused-argument # skipcq: PYL-W0613,PYL-R0201
-        """Return whether the include this build."""
+        """Return whether to include this build or not."""
         return True