+++ components/collector/tests/source_collectors/azure_devops/base.py
@@ -59,3 +59,61 @@ class AzureDevopsJobsTestCase(AzureDevopsTestCase):  # skipcq: PTC-W0046
                 build_status="failed",
             )
         ]
+
+
+class AzureDevopsPipelinesTestCase(AzureDevopsTestCase):  # skipcq: PTC-W0046
+    """Base class for Azure DevOps pipeline collectors."""
+
+    def setUp(self):
+        """Extend to set up pipeline data."""
+        super().setUp()
+        self.pipelines = dict(count=1, value=[dict(id=42, name="azure-pipelines.yml")])
+        self.pipeline_runs = dict(
+            count=3,
+            value=[
+                dict(
+                    state="completed",
+                    result="succeeded",
+                    createdDate="2019-10-15T12:20:10.1905868Z",
+                    finishedDate="2019-10-15T12:24:10.1905868Z",
+                    id=1,
+                    name="20191015.1",
+                    url=f"{self.url}/_build/results?buildId=1",
+                    _links=dict(web=dict(href=f"{self.url}/_build/results?buildId=1")),
+                ),
+                dict(
+                    state="completed",
+                    result="succeeded",
+                    createdDate="2019-10-15T12:30:10.1905868Z",
+                    finishedDate="2019-10-15T12:34:10.1905868Z",
+                    id=2,
+                    name="20191015.2",
+                    url=f"{self.url}/_build/results?buildId=2",
+                    _links=dict(web=dict(href=f"{self.url}/_build/results?buildId=2")),
+                ),
+                dict(
+                    state="inProgress",
+                    createdDate="2019-10-15T12:40:10.1905868Z",
+                    id=4,
+                    name="20191015.4",
+                    url=f"{self.url}/_build/results?buildId=4",
+                    _links=dict(web=dict(href=f"{self.url}/_build/results?buildId=4")),
+                )
+            ]
+        )
+        self.expected_entities = [
+            dict(
+                name="20191015.1",
+                key=f"{self.pipelines['value'][0]['id']}-20191015_1",  # safe_entity_key
+                url=f"{self.url}/_build/results?buildId=1",
+                build_date="2019-10-15",
+                build_status="completed",
+            ),
+            dict(
+                name="20191015.2",
+                key=f"{self.pipelines['value'][0]['id']}-20191015_2",  # safe_entity_key
+                url=f"{self.url}/_build/results?buildId=2",
+                build_date="2019-10-15",
+                build_status="completed",
+            )
+        ]