+++ components/collector/tests/source_collectors/jenkins/base.py
@@ -12,6 +12,6 @@ class JenkinsTestCase(SourceCollectorTestCase):  # skipcq: PTC-W0046
         """Extend to set up a Jenkins with a build."""
         super().setUp()
         self.set_source_parameter("failure_type", ["Failure"])
-        self.builds = [dict(result="FAILURE", timestamp="1552686540953")]
+        self.builds = [dict(result="FAILURE", timestamp=1552686540953)]
         self.job_url = "https://job"
         self.job2_url = "https://job2"