+++ components/collector/tests/source_collectors/jenkins/test_source_up_to_dateness.py
@@ -31,7 +31,7 @@ class JenkinsSourceUpToDatenessTest(JenkinsTestCase):
     async def test_ignore_failed_builds(self):
         """Test that failed builds can be ignored."""
         self.set_source_parameter("result_type", ["Success"])
-        self.builds.append(dict(result="SUCCESS", timestamp="1553686540953"))
+        self.builds.append(dict(result="SUCCESS", timestamp=1553686540953))
         jenkins_json = dict(jobs=[dict(name="job", url=self.job_url, buildable=True, color="red", builds=self.builds)])
         response = await self.collect(get_request_json_return_value=jenkins_json)
         expected_value = str((date.today() - date.fromisoformat("2019-03-27")).days)