+++ docs/src/changelog.md
@@ -29,6 +29,7 @@ If your currently installed *Quality-time* version is not v4.3.0, please read th
 - Make Jira issue identifiers uppercase when querying Jira for the status of issues, so that users don't have to enter uppercase Jira identifiers in *Quality-time*. Closes [#4450](https://github.com/ICTU/quality-time/issues/4450).
 - Allow for setting default labels on Jira issues created from *Quality-time*. The labels can be configured in the issue tracker tab under the report header. Closes [#4468](https://github.com/ICTU/quality-time/issues/4468).
 - Add a `target="_blank"` to URLs added in comments so that links open in a new tab by default. Can be prevented by adding an explicit `target="_top"` to URLs, for example: `This is an <a href="https://example.org" target="_top">example</a>`. Closes [#4521](https://github.com/ICTU/quality-time/issues/4521).
+- Metric "Job runs within time period", to keep track of deployment frequency in Jenkins, GitLab and/or Azure DevOps. Closes [#3498](https://github.com/ICTU/quality-time/issues/3498).
 
 ### Changed